requirements.txt
fastapi[standard]
uvicorn
passlib[bcrypt]
sqlalchemy
psycopg2-binary
alembic
python-decouple
python-jose
httpx
pytest
